modif posx posy to float, ajout var zoom

+++ includes/boss.class.hpp
@@ -8,7 +8,7 @@ public:
 
 	int name;
 
-	Boss( int x, int y, int status, int name );
+	Boss( float x, float y, int status, int name );
 	Boss( Boss const & src );
 	Boss & operator=( Boss const & rhs );
 	virtual ~Boss( void );
