+++ src/bomb.class.cpp
@@ -2,4 +2,4 @@
 
 Bomb::~Bomb( void ) {}
 
-Bomb::Bomb( int x, int y, int status ) : Entity( BOMB, 0, x, y, status ) {}
+Bomb::Bomb( float x, float y, int status ) : Entity( BOMB, 0, x, y, status ) {}
